Add support for siblings of the homepage
At least for the landing page itself, and the top menu-bar, add support
for the case of siblings of the homepage (i.e. pages in the same
directory as the homepage _index.md file; Hugo actually considers them
children of the homepage of type page, vs. directories which are section
children of the homepage).

Signed-off-by: Daniel F. Dickinson <cshored@thecshore.com>