Fix compatibility with demo theme with filter fix
This is involves another refactor of the local navigation bar.  It was
necessitated by the Hugo demo site exposing some features of Hugo that
broke the theme as it stood.  In particular page and section aliases
resulted in double headings and other strangeness.  So rather than add
a whole list of conditions in multiple places we take advantage of new
features of current Hugo that allow us turn the filter logic into a
partial we call from anywhere we need it.

While we're at it we finally get current-page detection working (using a
feature of Hugo that we just discovered, and another that was added
recently).

Signed-off-by: Daniel F. Dickinson <cshored@thecshore.com>