Add basic taxonomy support (but not links yet)
We get some basic taxonomy configuration into the layouts (rename
section.html to list.html so that it handles all list pages by
default (including taxonomy and taxonomyTerm pages).  In the process we
do some initial hacking on the sidebar/local-nav so that it doesn't
break the build with taxonomy enabled.
We also add taxonomic information to most of the content in the example
site.

Signed-off-by: Daniel F. Dickinson <cshored@thecshore.com>