Remove non-working accessibility nav links
The navigation bars that were supposed to be only visible to screen
readers fail to be visible to screen readers and not the visual display.
In addition most current screenreaders have reasonably handling of the
ARIA navigation features found on this site, and the navigation becomes
more of an annoyance than help, so we remove the 'non-visual-only'
navigation hack.

Signed-off-by: Daniel F. Dickinson <cshored@thecshore.com>