Workaround issue with .Site.Home.Pages
We implement a workaround for the fact that .Site.Home.Pages does not
include .Site.Home.Sections, unlike every other .Pages including
.Sections.  While we're at it make the local nav bar a ring at the level
below .Site.Home (i.e. all 'normal pages/sections' are included in the
circular traversal and the last page points to the first for Next, and
vice versa for Prev).

Signed-off-by: Daniel F. Dickinson <cshored@thecshore.com>