Update documentation, README, and About page
We're adding more documentation than comfortably fits on one page, so
reorganize that.  The README is now only for viewing from the repo
browser and the [Hugo Themes Showcase](themes.gohugo.io) so we make
tweaks for satisfying that usage (most was done in a previous commit).

Finally since we've improve the navigation, make the 'About' page a site
top-level page, sorted first, and have it as the landing / introduction
page.

Signed-off-by: Daniel F. Dickinson <cshored@thecshore.com>