Fix for "Top menubar not mobile-friendly"
We switch to a mobile-first design that starts with the menu bar items
taking up the entire width of the screen and stacked horizontally, and
moving as the page scrolls rather than 'fixed' or 'sticky'.  In addition
the sub-menus appear below the main menu item on the same z-index,
rather than popping up.  Sub-sub-menus are disabled.  On larger
screens the flyout menu is used.

Closes: #11

Signed-off-by: Daniel F. Dickinson <cshored@thecshore.com>