Fix for "Page ToC should not be on sidebar"
This commit Closes: #9 by moving around a number of screen elements.
The local-navbar and section-navbar now live outside of the 'main'
element, and their location depends on screen size.  On smaller screens,
the presumption is of a mobile device and so the navbar are above the
page title (below the top menubar).  On larger screens they become part
of the sidebar (in effect, not in HTML).

The ToC is moved into main article and floated up and right so that it
occupies a corner of the article.

Signed-off-by: Daniel F. Dickinson <cshored@thecshore.com>