Fix date-sorted lists to ignore no frontmatter date
Without a frontmatter date, the date was falling back to the git commit
date.  This resulted in pages with no date sorting ahead of pages with
a date older than the last time a page with no date was committed.
We would rather pages without a date not show in date-sorted lists such
as the sidebar and 'Recent Changes' or 'Recent News'.

Signed-off-by: Daniel F. Dickinson <cshored@thecshore.com>