Refactor header and use feature detection
We refactor the header to use a hamburger menu to make the top bar stay fixed at the top,
otherwise it scrolls away with the top of the page.  If the active area is selected then
the top bar is made fixed and appears at the top of the page.  In addition if the
user hovers over the hamburger icon the top bar will be fixed at the top of the page
while hover over the hamburger or top bar.

Some additional related styling changes are made to accomodate this.

Signed-off-by: Daniel F. Dickinson <cshored@thecshore.com>