Fix ToCSS Usage
This commit uses resources.ToCSS the way it was intended.
We now only execuate one file as a template (in order to pick up
params), and the rest of the css/scss is not in assets and gets
found by ToCSS via includePaths and a bunch of @import statements.

This has some benefits including meaningful line number and error messages,
Hugo's watch mechanism picks up changes and regenerates the target CSS
since assets is in the watch list, and we now fingerprint the assets so that
we do cache busting for the CSS.

While we were at it we moved normalize's virtual presence to assets/scss/normalize.
This will allow us to vendor the module because we are now targeting a directory
instead of a file.

Signed-off-by: Daniel F. Dickinson <cshored@thecshore.com>