Fix Missing Colour From Config
The attempt to deal with ToCSS in extended vs. basic Hugo broke
the use of colour (and other style variables exposed to config.toml
as params) params.  This fixes that and relies on Hugo 0.66.0 to
take care of the need for pregenerated resources (assuming the
resources directory has been generated).

Signed-off-by: Daniel F. Dickinson <cshored@thecshore.com>